requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9
Werkzeug>=3.0
requests>=2.31
pandas>=2.0
plotly>=5.0
python-dotenv>=1.0
